+++ src/main/java/com/entropyteam/entropay/common/Range.java
@@ -1,0 +1,5 @@
+package com.entropyteam.entropay.common;
+
+public class Range {
+    public String range;
+}
