+++ src/main/java/com/entropyteam/entropay/common/Sort.java
@@ -1,0 +1,5 @@
+package com.entropyteam.entropay.common;
+
+public class Sort {
+    public String sort;
+}
