+++ src/main/java/com/entropyteam/entropay/employees/clients/controllers/ClientController.java
@@ -6,9 +6,9 @@ import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import com.entropyteam.entropay.common.BaseController;
 import com.entropyteam.entropay.employees.clients.dtos.ClientDto;
 import com.entropyteam.entropay.employees.clients.services.ClientService;
-import com.entropyteam.entropay.employees.common.BaseController;
 
 @RestController
 @CrossOrigin
