+++ src/main/java/com/entropyteam/entropay/employees/clients/repositories/ClientRepository.java
@@ -1,8 +1,8 @@
 package com.entropyteam.entropay.employees.clients.repositories;
 
 import java.util.UUID;
+import com.entropyteam.entropay.common.BaseRepository;
 import com.entropyteam.entropay.employees.clients.models.Client;
-import com.entropyteam.entropay.employees.common.BaseRepository;
 
 public interface ClientRepository extends BaseRepository<Client, UUID> {
