+++ src/main/java/com/entropyteam/entropay/employees/clients/services/ClientService.java
@@ -3,11 +3,11 @@ package com.entropyteam.entropay.employees.clients.services;
 import java.util.UUID;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import com.entropyteam.entropay.common.BaseRepository;
+import com.entropyteam.entropay.common.BaseService;
 import com.entropyteam.entropay.employees.clients.dtos.ClientDto;
 import com.entropyteam.entropay.employees.clients.models.Client;
 import com.entropyteam.entropay.employees.clients.repositories.ClientRepository;
-import com.entropyteam.entropay.employees.common.BaseRepository;
-import com.entropyteam.entropay.employees.common.BaseService;
 
 import lombok.extern.log4j.Log4j2;
