--- src/main/java/com/entropyteam/entropay/employees/common/Range.java
@@ -1,5 +1,0 @@
-package com.entropyteam.entropay.employees.common;
-
-public class Range {
-    public String range;
-}
