--- src/main/java/com/entropyteam/entropay/employees/common/Sort.java
@@ -1,5 +1,0 @@
-package com.entropyteam.entropay.employees.common;
-
-public class Sort {
-    public String sort;
-}
