+++ src/main/java/com/entropyteam/entropay/employees/entropist/repositories/EmployeeRepository.java
@@ -2,7 +2,7 @@ package com.entropyteam.entropay.employees.entropist.repositories;
 
 import java.util.List;
 import java.util.UUID;
-import com.entropyteam.entropay.employees.common.BaseRepository;
+import com.entropyteam.entropay.common.BaseRepository;
 import com.entropyteam.entropay.employees.entropist.models.Employee;
 
 public interface EmployeeRepository extends BaseRepository<Employee, UUID> {
