+++ src/main/java/com/entropyteam/entropay/employees/entropist/services/EmployeeService.java
@@ -4,8 +4,8 @@ import java.util.Objects;
 import java.util.UUID;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import com.entropyteam.entropay.employees.common.BaseRepository;
-import com.entropyteam.entropay.employees.common.BaseService;
+import com.entropyteam.entropay.common.BaseRepository;
+import com.entropyteam.entropay.common.BaseService;
 import com.entropyteam.entropay.employees.entropist.dtos.EmployeeDto;
 import com.entropyteam.entropay.employees.entropist.models.Employee;
 import com.entropyteam.entropay.employees.entropist.repositories.EmployeeRepository;
