+++ src/main/java/com/entropyteam/entropay/employees/project/models/ProjectType.java
@@ -3,7 +3,7 @@ package com.entropyteam.entropay.employees.project.models;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
-import com.entropyteam.entropay.employees.common.BaseEntity;
+import com.entropyteam.entropay.common.BaseEntity;
 import com.entropyteam.entropay.employees.project.dtos.ProjectTypeDto;
 
 @Entity(name = "ProjectType")
