+++ src/main/java/com/entropyteam/entropay/employees/project/repositories/ProjectRepository.java
@@ -1,7 +1,7 @@
 package com.entropyteam.entropay.employees.project.repositories;
 
 import java.util.UUID;
-import com.entropyteam.entropay.employees.common.BaseRepository;
+import com.entropyteam.entropay.common.BaseRepository;
 import com.entropyteam.entropay.employees.project.models.Project;
 
 public interface ProjectRepository extends BaseRepository<Project, UUID> {
