+++ src/main/java/com/entropyteam/entropay/employees/project/repositories/ProjectTypeRepository.java
@@ -1,7 +1,7 @@
 package com.entropyteam.entropay.employees.project.repositories;
 
 import java.util.UUID;
-import com.entropyteam.entropay.employees.common.BaseRepository;
+import com.entropyteam.entropay.common.BaseRepository;
 import com.entropyteam.entropay.employees.project.models.ProjectType;
 
 public interface ProjectTypeRepository extends BaseRepository<ProjectType, UUID> {
