+++ src/main/java/com/entropyteam/entropay/employees/project/services/ProjectService.java
@@ -4,10 +4,10 @@ import java.util.UUID;
 import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import com.entropyteam.entropay.common.BaseRepository;
+import com.entropyteam.entropay.common.BaseService;
 import com.entropyteam.entropay.employees.clients.models.Client;
 import com.entropyteam.entropay.employees.clients.repositories.ClientRepository;
-import com.entropyteam.entropay.employees.common.BaseRepository;
-import com.entropyteam.entropay.employees.common.BaseService;
 import com.entropyteam.entropay.employees.project.dtos.ProjectDto;
 import com.entropyteam.entropay.employees.project.models.Project;
 import com.entropyteam.entropay.employees.project.models.ProjectType;
