+++ src/main/java/com/entropyteam/entropay/employees/project/services/ProjectTypeService.java
@@ -4,8 +4,8 @@ import java.util.Objects;
 import java.util.UUID;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import com.entropyteam.entropay.employees.common.BaseRepository;
-import com.entropyteam.entropay.employees.common.BaseService;
+import com.entropyteam.entropay.common.BaseRepository;
+import com.entropyteam.entropay.common.BaseService;
 import com.entropyteam.entropay.employees.project.dtos.ProjectTypeDto;
 import com.entropyteam.entropay.employees.project.models.ProjectType;
 import com.entropyteam.entropay.employees.project.repositories.ProjectTypeRepository;
